+++ swinTransformer/templates/work_done_info.html
@@ -1,3 +1,4 @@
 {{ user_id }}:{{ username }}
 您最近使用了我们的网站的分割功能
 快去看看你的最新结果吧🤠
+<img alt="done" src={{ image }}>
